Changed TimeStamp to LocalDateTime Create Factories for testing Replace .findById with getById Replace (and remove) InvalidArgumentException with ExpertConsultationException Adjusted testing to new modifications
