--- src/main/java/ro/code4/expertconsultation/exception/InvalidArgumentException.java
@@ -1,8 +1,0 @@
-package ro.code4.expertconsultation.exception;
-
-public class InvalidArgumentException extends RuntimeException{
-
-    public InvalidArgumentException(String message) {
-        super(message);
-    }
-}
